moved persistentstorage structs in a separate file and added new tests
