+++ src/utils.rs
@@ -1,0 +1,7 @@
+use std::time::{SystemTime, UNIX_EPOCH};
+pub fn get_epoch_ms() -> u128 {
+    SystemTime::now()
+        .duration_since(UNIX_EPOCH)
+        .unwrap()
+        .as_millis()
+}
